pyproject.toml
[project]
name = "microservicesfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "redis-om>=0.3.5",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
